+++ cspy/checks.py
@@ -70,9 +70,7 @@ def _check_path(G, max_res, min_res, direction, algorithm):
         if not has_path(G, 'Source', 'Sink'):
             raise NetworkXException("Disconnected Graph")
         if negative_edge_cycle(G):
-            raise NetworkXUnbounded
-    except NetworkXUnbounded:
-        raise Exception("A negative cost cycle was found.")
+            raise NetworkXException("A negative cost cycle was found.")
     except NetworkXException as e:
         raise Exception("An error occurred: {}".format(e))