+++ cspy/preprocessing.py
@@ -1,9 +1,9 @@
-import logging
-import networkx as nx
+from logging import getLogger
+from networkx import single_source_bellman_ford
 from cspy.checks import (_check_res, _check_direction, _check_graph_attr,
                          _check_edge_attr, _check_path, _check_REFs)
 
-log = logging.getLogger(__name__)
+log = getLogger(__name__)
 
 
 def check(G, max_res=None, min_res=None, direction=None, algorithm=None):
@@ -36,13 +36,15 @@ def check(G, max_res=None, min_res=None, direction=None, algorithm=None):
 
 def prune_graph(G, max_res, min_res):
     """Removes nodes that cannot be reached due to resource limits.
-    Note:
-        path_s and path_t contains all partial paths e.g.
-            {node_i: [Source, ..., node_k, node_i]}.
-        or  {node_i: [Sink, ..., node_k, node_i]}
-        Hence, if node_i is found to violate a resource bound, it is
-        because of node_k, therefore, we add path[key][-2] = node_k to
-        the dictionary of nodes to remove.
+
+    Note
+    -----
+    path_s and path_t contains all partial paths e.g.
+        {node_i: [Source, ..., node_k, node_i]}.
+    or  {node_i: [Sink, ..., node_k, node_i]}
+    Hence, if node_i is found to violate a resource bound, it is
+    because of node_k, therefore, we add path[key][-2] = node_k to
+    the dictionary of nodes to remove.
     """
 
     def _check_resource(r):
@@ -53,12 +55,12 @@ def prune_graph(G, max_res, min_res):
             return attr_dict['res_cost'][r]
 
         # Get paths from source to all other nodes
-        length_s, path_s = nx.single_source_bellman_ford(G,
-                                                         'Source',
-                                                         weight=__get_weight)
-        length_t, path_t = nx.single_source_bellman_ford(G.reverse(copy=True),
-                                                         'Sink',
-                                                         weight=__get_weight)
+        length_s, path_s = single_source_bellman_ford(G,
+                                                      'Source',
+                                                      weight=__get_weight)
+        length_t, path_t = single_source_bellman_ford(G.reverse(copy=True),
+                                                      'Sink',
+                                                      weight=__get_weight)
         try:
             # Collect nodes in paths that violate the resource bounds
             # see note above